SIG-517 Moved apptimize and flex horeca email logic to Django app email integrations.
