+++ api/app/signals/apps/signals/tasks.py
@@ -37,60 +37,6 @@ def push_to_sigmax(pk: int):
 
 
 @app.task
-def send_mail_flex_horeca(pk):
-    """Send email to Flex Horeca Team when applicable.
-
-    :param pk: Signal object id
-    :returns:
-    """
-    try:
-        signal = Signal.objects.get(id=pk)
-    except Signal.DoesNotExist as e:
-        log.exception(str(e))
-        return
-
-    if _is_signal_applicable_for_flex_horeca(signal):
-        template = loader.get_template('mail_flex_horeca.txt')
-        context = {'signal': signal, }
-        message = template.render(context)
-        send_mail(
-            subject='Nieuwe melding op meldingen.amsterdam.nl',
-            message=message,
-            from_email=settings.NOREPLY,
-            recipient_list=(settings.EMAIL_FLEX_HORECA_INTEGRATION_ADDRESS,),
-            fail_silently=False)
-
-
-def _is_signal_applicable_for_flex_horeca(signal):
-    """Is given `Signal` applicable for Flex Horeca Team.
-
-    Flex Horeca Team can't check the Signals Dashboard on friday and
-    saterday. That's why we send them an e-mail notification on these days
-    for new `Signal` objects that are created.
-
-    :param signal: Signal object
-    :returns: bool
-    """
-    today = timezone.now()
-    weekday = today.isoweekday()
-    is_friday_or_saterday = weekday == 5 or weekday == 6
-    if not is_friday_or_saterday:
-        return False
-
-    eligible_main_categories = 'Overlast Bedrijven en Horeca'
-    eligible_sub_categories = (
-        'Geluidsoverlast muziek',
-        'Geluidsoverlast installaties',
-        'Overlast terrassen',
-        'Stankoverlast')
-    is_applicable_for_flex_horeca = (
-            signal.category.main == eligible_main_categories
-            and signal.category.sub in eligible_sub_categories)
-
-    return is_applicable_for_flex_horeca
-
-
-@app.task
 def task_save_csv_files_datawarehouse():
     """Celery task to save CSV files for Datawarehouse.
